+++ nagios-plugin-check_raid/ChangeLog.md
@@ -1,7 +1,15 @@
 ## Unreleased
 
+- arcconf: dead disks have no id, use physical location instead. [#90][]
+- sudo: detect if sudo has `-A` option [#88][]
+
+## 3.2.2 (2014-11-15)
+
 - cciss: fix parsing enclosure with no enclosure name [#84][]
 - megacli: actually report that cache is disabled. [#85][]
+- arcconf: accept 'Ready' as OK drive state. [#87][]
+- tw_cli: tweak VERIFYING state [#89][]
+- megacli: JBOD state of physical device is OK as well [#82][]
 
 ## 3.2.1 (2014-10-07)
 
@@ -129,3 +137,8 @@ NOTES:
 [#65]: https://github.com/glensc/nagios-plugin-check_raid/issues/65
 [#84]: https://github.com/glensc/nagios-plugin-check_raid/issues/84
 [#85]: https://github.com/glensc/nagios-plugin-check_raid/issues/85
+[#87]: https://github.com/glensc/nagios-plugin-check_raid/issues/87
+[#89]: https://github.com/glensc/nagios-plugin-check_raid/pull/89
+[#82]: https://github.com/glensc/nagios-plugin-check_raid/pull/82
+[#90]: https://github.com/glensc/nagios-plugin-check_raid/issues/90
+[#88]: https://github.com/glensc/nagios-plugin-check_raid/issues/88